// srcs/front/src/types/apiResponses.ts
export interface UserRegister {
    username: string;
    email: string;
    password: string;
    first_name: string;
    last_name: string;
}






export interface UserProfilePicturesResponse {
    is_profile_picture: boolean;
    url: string;
}

export interface UserProfileInfosResponse {
    biography: string;
    fame_rating: number;
}

export interface UserProfileAttrsResponse {
    first_name: string;
    gender: string;
    last_name: string;
    sexual_preference: string;
    user_id: number;
    username: string;
    email?: string;
    location?: string;
    birthdate?: string;
}


export interface UserProfileInteractionsResponse {
    is_same: boolean;
    likes_count: number;
    views_count: number;
    is_connected?: number;
    interaction_status?: string;
    conversation_id?: number;
}

export interface UserProfileResponse {
  interactions: UserProfileInteractionsResponse;
  interests: string[];
  pictures: UserProfilePicturesResponse[];
  profile: UserProfileInfosResponse;
  user: UserProfileAttrsResponse
}

export interface NotificationsResponse {
    created_at: string;
    is_read: boolean;
    notification_id: number;
    picture_url: UserProfilePicturesResponse[];
    source_user_id: number;
    type: string;
    user_id: number;
    username: string;
}

export interface MessagesResponse {
    content: string;
    id: number;
    is_read: boolean;
    sender_id: number;
    sent_at: string;
}

export interface ConversationsResponse {
    conversation_id: number;
    last_active_at: string;
    first_name: string;
    last_name: string;
    last_online: string;
    peer_id: number;
    profile_picture_url: UserProfilePicturesResponse[];
    username: string;
    messages_list?: MessagesResponse[];
}

export interface SuggestionsResponse {
    user_id: number;
    fame_rating: number;
    age: string;
    last_name: string;
    first_name: string;
    username: string;
    location: string;
    is_liked: number;
    profile_picture_url: UserProfilePicturesResponse[];
}


export interface UserLocation {
    id: number;
    username: string;
    latitude: number;
    longitude: number;
    profile_picture_url: UserProfilePicturesResponse[];

}

export interface Location {
    latitude: number,
    longitude: number
}

export interface BlocksResponse {
    blocked_id: number,
    username: string,
    first_name: string,
    last_name: string,
    profile_picture_url: UserProfilePicturesResponse[]
    unblock?: boolean
}

export interface LikesResponse {
    liker_id: number,
    username: string,
    first_name: string,
    last_name: string,
    profile_picture_url: UserProfilePicturesResponse[],
    liked_back: number,
    unlike?: boolean
}

export interface ViewsListResponse {
    viewer_id: number,
    username: string,
    first_name: string,
    last_name: string,
    profile_picture_url: UserProfilePicturesResponse[],
    viewed_at: any
}


export interface UserDatesResponse {
    id: number,
    proposer_id: number,
    partner_id: number,
    proposer_username: string,
    partner_username: string,
    proposer_avatar: UserProfilePicturesResponse[],
    partner_avatar: UserProfilePicturesResponse[],
    location: string,
    scheduled_at: any,
    description: string,
    status: string,
}